requirements.txt
python=3.10
Pillow~=9.2.0
